added types to "app". move inference classes into own file.
